+++ src/fxr.ts
@@ -1587,60 +1587,31 @@ enum ResourceType {
  */
 export enum ScaleCondition {
   /**
-   * Always scale the property, unless the scaling mode is {@link ScalingMode.InstancesOnly}
-   * or {@link ScalingMode.ParticleModifierOnly}.
+   * Always scale the property, unless {@link ScaleOptions.scaleStructural scaleStructural}
+   * is *disabled*.
    */
   True = 1,
   /**
-   * Only scale the property if the value is not -1.
+   * Only scale the property if {@link ScaleOptions.scaleViewDistance scaleViewDistance}
+   * and {@link ScaleOptions.scaleStructural scaleStructural} are *enabled*.
    */
-  IfNotMinusOne = 2,
+  Distance = 2,
   /**
-   * Only scale the property if view distance properties are being scaled.
+   * Only scale the property if the value is not -1 and
+   * {@link ScaleOptions.scaleViewDistance scaleViewDistance} and
+   * {@link ScaleOptions.scaleStructural scaleStructural} are *enabled*.
    */
-  Distance = 3,
+  DistanceIfNotMinusOne = 3,
   /**
-   * Only scale the property if view distance properties are being scaled and the value is not -1.
+   * Only scale the property if {@link ScaleOptions.scaleParticleModifier scaleParticleModifier}
+   * is *disabled*.
    */
-  DistanceIfNotMinusOne = 4,
+  InstanceSize = 4,
   /**
-   * Always scale the property, unless the scaling mode is {@link ScalingMode.ParticleModifierOnly}.
+   * Only scale the property if {@link ScaleOptions.scaleParticleModifier scaleParticleModifier}
+   * is *enabled*.
    */
-  InstanceSize = 5,
-  /**
-   * Only scale the property if the scaling mode is {@link ScalingMode.ParticleModifierOnly}.
-   */
-  ParticleModifier = 6,
-}
-
-/**
- * Controls what properties are scaled by the `scale` methods for nodes,
- * configs, and actions.
- */
-export enum ScalingMode {
-  /**
-   * Scale all scaling properties.
-   */
-  All = 0,
-  /**
-   * Scale all non-view distance-based scaling properties.
-   */
-  NoViewDistance = 1,
-  /**
-   * Only scale properties that control the size of instances.
-   * 
-   * "Instances" here refer to particles and certain other visual effects, like
-   * light sources. This option will not change the size of emitters, or any
-   * translations, only individual appearance instances.
-   */
-  InstancesOnly = 2,
-  /**
-   * Only scale the scale properties in {@link ParticleModifier} actions.
-   * 
-   * Note: The {@link ParticleModifier.prototype.speed speed} property will
-   * *not* be scaled.
-   */
-  ParticleModifierOnly = 3,
+  ParticleModifier = 5,
 }
 
 /**
@@ -1854,7 +1825,7 @@ export interface IConfig {
    * Scale the config by the given `factor`. This can be used to change
    * the size of effect created by the config.
    */
-  scale(factor: number, options?: { mode?: ScalingMode }): this
+  scale(factor: number, options?: ScaleOptions): this
 
   toJSON(): any
   serialize(options?: FXRSerializeOptions): any
@@ -1880,6 +1851,35 @@ export interface NodeColorOptions {
   activeState?: number
   side?: 'start' | 'end' | 'middle'
   layer?: number
+}
+
+export interface ScaleOptions {
+  /**
+   * Scale the scale properties in {@link ParticleModifier} actions _instead
+   * of_ those in the appearance actions.
+   * 
+   * Defaults to `false`.
+   */
+  scaleParticleModifier?: boolean
+  /**
+   * Scale properties that are based on the distance to the camera, such as the
+   * min/max distance for many apperance actions.
+   * 
+   * Defaults to `false`.
+   */
+  scaleViewDistance?: boolean
+  /**
+   * Scale properties that control the size of the structure of the effect,
+   * i.e. ones that don't affect the size of the appearance instances, except
+   * the emitter size for GPU particle actions.
+   * 
+   * Defaults to `true`.
+   */
+  scaleStructural?: boolean
+}
+
+const defaultScaleOptions: ScaleOptions = {
+  scaleStructural: true,
 }
 
 export type AnyAction = Action | DataAction
@@ -2725,9 +2725,9 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: true,
     props: {
       speed: { default: 0, scale: 1, time: 1 },
-      scaleX: { default: 1, scale: 6 },
-      scaleY: { default: 1, scale: 6 },
-      scaleZ: { default: 1, scale: 6 },
+      scaleX: { default: 1, scale: 5 },
+      scaleY: { default: 1, scale: 5 },
+      scaleZ: { default: 1, scale: 5 },
       color: { default: [1, 1, 1, 1], color: 2 },
       uniformScale: { default: false, field: 0 },
     },
@@ -2763,11 +2763,11 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: true,
     props: {
       duration: { default: -1, time: 3 },
-      threshold0: { default: 10000, field: 1, scale: 3 },
-      threshold1: { default: 10000, field: 1, scale: 3 },
-      threshold2: { default: 10000, field: 1, scale: 3 },
-      threshold3: { default: 10000, field: 1, scale: 3 },
-      threshold4: { default: 10000, field: 1, scale: 3 },
+      threshold0: { default: 10000, field: 1, scale: 2 },
+      threshold1: { default: 10000, field: 1, scale: 2 },
+      threshold2: { default: 10000, field: 1, scale: 2 },
+      threshold3: { default: 10000, field: 1, scale: 2 },
+      threshold4: { default: 10000, field: 1, scale: 2 },
       unk_ac6_f1_5: { default: 0, field: 1 },
     },
     games: {
@@ -3061,7 +3061,7 @@ const ActionData: Record<string, ActionDataEntry> = {
     props: {
       texture: { default: 1, field: 1, resource: 0, textureType: 'a' },
       blendMode: { default: BlendMode.Normal, field: 1 },
-      size: { default: 1, scale: 5 },
+      size: { default: 1, scale: 4 },
       color1: { default: [1, 1, 1, 1], color: 2 },
       color2: { default: [1, 1, 1, 1], color: 2 },
       color3: { default: [1, 1, 1, 1], color: 1 },
@@ -3088,22 +3088,22 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 0, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
-      unk_ds3_f2_29: { default: 0, field: 2, scale: 3 },
+      unk_ds3_f2_29: { default: 0, field: 2, scale: 2 },
       unk_sdt_f2_30: { default: 0, field: 2 },
       unk_sdt_f2_31: { default: 0, field: 1 },
       unk_sdt_f2_32: { default: false, field: 0 },
@@ -3143,7 +3143,7 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: false,
     props: {
       blendMode: { default: BlendMode.Normal, field: 1 },
-      length: { default: 1, scale: 5 },
+      length: { default: 1, scale: 4 },
       color1: { default: [1, 1, 1, 1], color: 1 },
       color2: { default: [1, 1, 1, 1], color: 2 },
       startColor: { default: [1, 1, 1, 1], color: 2 },
@@ -3171,22 +3171,22 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 0, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
-      unk_ds3_f2_29: { default: 0, field: 2, scale: 3 },
+      unk_ds3_f2_29: { default: 0, field: 2, scale: 2 },
       unk_sdt_f2_30: { default: 0, field: 1 },
       unkHideIndoors: { default: 0, field: 1 },
       unk_sdt_f2_32: { default: false, field: 0 },
@@ -3226,8 +3226,8 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: false,
     props: {
       blendMode: { default: BlendMode.Normal, field: 1 },
-      width: { default: 1, scale: 5 },
-      length: { default: 1, scale: 5 },
+      width: { default: 1, scale: 4 },
+      length: { default: 1, scale: 4 },
       color1: { default: [1, 1, 1, 1], color: 1 },
       color2: { default: [1, 1, 1, 1], color: 2 },
       startColor: { default: [1, 1, 1, 1], color: 2 },
@@ -3256,22 +3256,22 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 0, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
-      unk_ds3_f2_29: { default: 0, field: 2, scale: 3 },
+      unk_ds3_f2_29: { default: 0, field: 2, scale: 2 },
       unk_sdt_f2_30: { default: 0, field: 2 },
       unk_sdt_f2_31: { default: 0, field: 1 },
       unk_sdt_f2_32: { default: false, field: 0 },
@@ -3315,8 +3315,8 @@ const ActionData: Record<string, ActionDataEntry> = {
       offsetX: { default: 0 },
       offsetY: { default: 0 },
       offsetZ: { default: 0 },
-      width: { default: 1, scale: 5 },
-      height: { default: 1, scale: 5 },
+      width: { default: 1, scale: 4 },
+      height: { default: 1, scale: 4 },
       color1: { default: [1, 1, 1, 1], color: 2 },
       color2: { default: [1, 1, 1, 1], color: 2 },
       color3: { default: [1, 1, 1, 1], color: 1 },
@@ -3371,22 +3371,22 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
-      unk_ds3_f2_29: { default: 0, field: 2, scale: 3 },
+      unk_ds3_f2_29: { default: 0, field: 2, scale: 2 },
       shadowDarkness: { default: 0, field: 2 },
       unkHideIndoors: { default: 0, field: 1 },
       unk_sdt_f2_32: { default: false, field: 0 },
@@ -3435,8 +3435,8 @@ const ActionData: Record<string, ActionDataEntry> = {
       offsetX: { default: 0 },
       offsetY: { default: 0 },
       offsetZ: { default: 0 },
-      width: { default: 1, scale: 5 },
-      height: { default: 1, scale: 5 },
+      width: { default: 1, scale: 4 },
+      height: { default: 1, scale: 4 },
       rotationX: { default: 0 },
       rotationY: { default: 0 },
       rotationZ: { default: 0 },
@@ -3508,22 +3508,22 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
-      unk_ds3_f2_29: { default: 0, field: 2, scale: 3 },
+      unk_ds3_f2_29: { default: 0, field: 2, scale: 2 },
       shadowDarkness: { default: 0, field: 2 },
       unk_sdt_f2_31: { default: 0, field: 1 },
       unk_sdt_f2_32: { default: false, field: 0 },
@@ -3575,9 +3575,9 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: false,
     props: {
       model: { default: 80201, field: 1, resource: 1 },
-      sizeX: { default: 1, scale: 5 },
-      sizeY: { default: 1, scale: 5 },
-      sizeZ: { default: 1, scale: 5 },
+      sizeX: { default: 1, scale: 4 },
+      sizeY: { default: 1, scale: 4 },
+      sizeZ: { default: 1, scale: 4 },
       rotationX: { default: 0 },
       rotationY: { default: 0 },
       rotationZ: { default: 0 },
@@ -3640,19 +3640,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 2 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 2 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_26: { default: 1, field: 1 },
       unk_ds3_f2_27: { default: 0, field: 1 },
       unk_sdt_f2_29: { default: 0, field: 2 },
@@ -3701,7 +3701,7 @@ const ActionData: Record<string, ActionDataEntry> = {
     props: {
       texture: { default: 1, field: 1, resource: 0, textureType: 'a' },
       blendMode: { default: BlendMode.Normal, field: 1 },
-      width: { default: 1, scale: 5 },
+      width: { default: 1, scale: 4 },
       widthMultiplier: { default: 1 },
       startFadeEndpoint: { default: 0 },
       endFadeEndpoint: { default: 0 },
@@ -3725,7 +3725,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       orientation: { default: TracerOrientationMode.LocalZ, field: 1 },
       normalMap: { default: 0, field: 1, resource: 0, textureType: 'n' },
       segmentInterval: { default: 0, field: 2, time: 2 },
-      segmentDuration: { default: 1, field: 2, scale: 5, time: 5 },
+      segmentDuration: { default: 1, field: 2, scale: 4, time: 5 },
       concurrentSegments: { default: 100, field: 1 },
       segmentSubdivision: { default: 0, field: 1 },
       unk_ds3_f1_8: { default: 0, field: 2 },
@@ -3750,22 +3750,22 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
-      unk_ds3_f2_29: { default: 0, field: 2, scale: 3 },
+      unk_ds3_f2_29: { default: 0, field: 2, scale: 2 },
       shadowDarkness: { default: 0, field: 2 },
       unk_sdt_f2_31: { default: 0, field: 1 },
       unk_sdt_f2_32: { default: false, field: 0 },
@@ -3808,9 +3808,9 @@ const ActionData: Record<string, ActionDataEntry> = {
       offsetX: { default: 0 },
       offsetY: { default: 0 },
       offsetZ: { default: 0 },
-      sizeX: { default: 1, scale: 5 },
-      sizeY: { default: 1, scale: 5 },
-      sizeZ: { default: 1, scale: 5 },
+      sizeX: { default: 1, scale: 4 },
+      sizeY: { default: 1, scale: 4 },
+      sizeZ: { default: 1, scale: 4 },
       color: { default: [1, 1, 1, 1], color: 2 },
       unk_ds3_p1_7: { default: [1, 1, 1, 1] },
       intensity: { default: 1 },
@@ -3855,19 +3855,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 1, field: 2 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
       unk_ds3_f2_29: { default: 0, field: 1 },
@@ -3879,7 +3879,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_sdt_f2_35: { default: -1, field: 1 },
       unk_sdt_f2_36: { default: -2, field: 1 },
       unk_sdt_f2_37: { default: 0, field: 1 },
-      unk_sdt_f2_38: { default: 0, field: 2, scale: 3 },
+      unk_sdt_f2_38: { default: 0, field: 2, scale: 2 },
     },
     games: {
       [Game.DarkSouls3]: {
@@ -3913,8 +3913,8 @@ const ActionData: Record<string, ActionDataEntry> = {
       offsetX: { default: 0 },
       offsetY: { default: 0 },
       offsetZ: { default: 0 },
-      width: { default: 1, scale: 5 },
-      height: { default: 1, scale: 5 },
+      width: { default: 1, scale: 4 },
+      height: { default: 1, scale: 4 },
       color: { default: [1, 1, 1, 1], color: 2 },
       unk_ds3_p1_6: { default: [1, 1, 1, 1] },
       blurRadius: { default: 0.5 },
@@ -3941,19 +3941,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0.5, field: 2 },
       unk_ds3_f2_21: { default: 1, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 2 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
       unk_ds3_f2_29: { default: 0, field: 2 },
@@ -3988,7 +3988,7 @@ const ActionData: Record<string, ActionDataEntry> = {
     props: {
       diffuseColor: { default: [1, 1, 1, 1], color: 1 },
       specularColor: { default: [1, 1, 1, 1], color: 1 },
-      radius: { default: 10, scale: 5 },
+      radius: { default: 10, scale: 4 },
       unk_ds3_p1_3: { default: 0 },
       unk_ds3_p1_4: { default: 0 },
       unk_ds3_p1_5: { default: 0 },
@@ -4026,7 +4026,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 100, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
-      maxViewDistance: { default: 0, field: 2, scale: 3 },
+      maxViewDistance: { default: 0, field: 2, scale: 2 },
       volumeDensity: { default: 0, field: 2 },
       unk_sdt_f2_25: { default: 0, field: 2 },
       phaseFunction: { default: true, field: 0 },
@@ -4188,8 +4188,8 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_p1_7: { default: 0 },
       unk_ds3_p1_8: { default: 0 },
       particleAngularAccelerationZ: { default: 0, time: 4 },
-      particleGrowthRateX: { default: 0, scale: 5, time: 1 },
-      particleGrowthRateY: { default: 0, scale: 5, time: 1 },
+      particleGrowthRateX: { default: 0, scale: 4, time: 1 },
+      particleGrowthRateY: { default: 0, scale: 4, time: 1 },
       unk_ds3_p1_12: { default: 0 },
       color: { default: [1, 1, 1, 1], color: 1 },
       unk_ds3_p1_14: { default: 1 },
@@ -4249,29 +4249,29 @@ const ActionData: Record<string, ActionDataEntry> = {
       particleAngularAccelerationMin: { default: [0, 0, 0], field: 4, time: 4 },
       particleAngularAccelerationMax: { default: [0, 0, 0], field: 4, time: 4 },
       particleUniformScale: { default: false, field: 0 },
-      particleSizeX: { default: 1, field: 2, scale: 5 },
-      particleSizeY: { default: 1, field: 2, scale: 5 },
+      particleSizeX: { default: 1, field: 2, scale: 4 },
+      particleSizeY: { default: 1, field: 2, scale: 4 },
       unk_ds3_f1_73: { default: 1, field: 2 },
-      particleSizeXMin: { default: 0, field: 2, scale: 5 },
-      particleSizeYMin: { default: 0, field: 2, scale: 5 },
+      particleSizeXMin: { default: 0, field: 2, scale: 4 },
+      particleSizeYMin: { default: 0, field: 2, scale: 4 },
       unk_ds3_f1_76: { default: 0, field: 2 },
-      particleSizeXMax: { default: 0, field: 2, scale: 5 },
-      particleSizeYMax: { default: 0, field: 2, scale: 5 },
+      particleSizeXMax: { default: 0, field: 2, scale: 4 },
+      particleSizeYMax: { default: 0, field: 2, scale: 4 },
       unk_ds3_f1_79: { default: 0, field: 2 },
-      particleGrowthRateXStatic: { default: 0, field: 2, scale: 5, time: 1 },
-      particleGrowthRateYStatic: { default: 0, field: 2, scale: 5, time: 1 },
+      particleGrowthRateXStatic: { default: 0, field: 2, scale: 4, time: 1 },
+      particleGrowthRateYStatic: { default: 0, field: 2, scale: 4, time: 1 },
       unk_ds3_f1_82: { default: 0, field: 2 },
-      particleGrowthRateXMin: { default: 0, field: 2, scale: 5, time: 1 },
-      particleGrowthRateYMin: { default: 0, field: 2, scale: 5, time: 1 },
+      particleGrowthRateXMin: { default: 0, field: 2, scale: 4, time: 1 },
+      particleGrowthRateYMin: { default: 0, field: 2, scale: 4, time: 1 },
       unk_ds3_f1_85: { default: 0, field: 2 },
-      particleGrowthRateXMax: { default: 0, field: 2, scale: 5, time: 1 },
-      particleGrowthRateYMax: { default: 0, field: 2, scale: 5, time: 1 },
+      particleGrowthRateXMax: { default: 0, field: 2, scale: 4, time: 1 },
+      particleGrowthRateYMax: { default: 0, field: 2, scale: 4, time: 1 },
       unk_ds3_f1_88: { default: 0, field: 2 },
-      particleGrowthAccelerationXMin: { default: 0, field: 2, scale: 5, time: 4 },
-      particleGrowthAccelerationYMin: { default: 0, field: 2, scale: 5, time: 4 },
+      particleGrowthAccelerationXMin: { default: 0, field: 2, scale: 4, time: 4 },
+      particleGrowthAccelerationYMin: { default: 0, field: 2, scale: 4, time: 4 },
       unk_ds3_f1_91: { default: 0, field: 2 },
-      particleGrowthAccelerationXMax: { default: 0, field: 2, scale: 5, time: 4 },
-      particleGrowthAccelerationYMax: { default: 0, field: 2, scale: 5, time: 4 },
+      particleGrowthAccelerationXMax: { default: 0, field: 2, scale: 4, time: 4 },
+      particleGrowthAccelerationYMax: { default: 0, field: 2, scale: 4, time: 4 },
       unk_ds3_f1_94: { default: 0, field: 2 },
       rgbMultiplier: { default: 1, field: 2 },
       alphaMultiplier: { default: 1, field: 2 },
@@ -4322,7 +4322,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       particleRandomTurnIntervalMax: { default: 1, field: 1, time: 2 },
       traceParticles: { default: false, field: 0 },
       unk_ds3_f1_149: { default: 1, field: 2 },
-      particleTraceLength: { default: 1, field: 2, scale: 5, time: 2 },
+      particleTraceLength: { default: 1, field: 2, scale: 4, time: 2 },
       traceParticlesThreshold: { default: 0, field: 2 },
       traceParticleHead: { default: false, field: 0 },
       unk_ds3_f1_153: { default: 0, field: 1 },
@@ -4352,19 +4352,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
       unk_sdt_f2_29: { default: 0, field: 2 },
@@ -4416,8 +4416,8 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_p1_7: { default: 0 },
       unk_ds3_p1_8: { default: 0 },
       particleAngularAccelerationZ: { default: 0, time: 4 },
-      particleGrowthRateX: { default: 0, scale: 5, time: 1 },
-      particleGrowthRateY: { default: 0, scale: 5, time: 1 },
+      particleGrowthRateX: { default: 0, scale: 4, time: 1 },
+      particleGrowthRateY: { default: 0, scale: 4, time: 1 },
       unk_ds3_p1_12: { default: 0 },
       color: { default: [1, 1, 1, 1], color: 1 },
       unk_ds3_p1_14: { default: 1 },
@@ -4477,29 +4477,29 @@ const ActionData: Record<string, ActionDataEntry> = {
       particleAngularAccelerationMin: { default: [0, 0, 0], field: 4, time: 4 },
       particleAngularAccelerationMax: { default: [0, 0, 0], field: 4, time: 4 },
       particleUniformScale: { default: false, field: 0 },
-      particleSizeX: { default: 1, field: 2, scale: 5 },
-      particleSizeY: { default: 1, field: 2, scale: 5 },
+      particleSizeX: { default: 1, field: 2, scale: 4 },
+      particleSizeY: { default: 1, field: 2, scale: 4 },
       unk_ds3_f1_73: { default: 1, field: 2 },
-      particleSizeXMin: { default: 0, field: 2, scale: 5 },
-      particleSizeYMin: { default: 0, field: 2, scale: 5 },
+      particleSizeXMin: { default: 0, field: 2, scale: 4 },
+      particleSizeYMin: { default: 0, field: 2, scale: 4 },
       unk_ds3_f1_76: { default: 0, field: 2 },
-      particleSizeXMax: { default: 0, field: 2, scale: 5 },
-      particleSizeYMax: { default: 0, field: 2, scale: 5 },
+      particleSizeXMax: { default: 0, field: 2, scale: 4 },
+      particleSizeYMax: { default: 0, field: 2, scale: 4 },
       unk_ds3_f1_79: { default: 0, field: 2 },
-      particleGrowthRateXStatic: { default: 0, field: 2, scale: 5, time: 1 },
-      particleGrowthRateYStatic: { default: 0, field: 2, scale: 5, time: 1 },
+      particleGrowthRateXStatic: { default: 0, field: 2, scale: 4, time: 1 },
+      particleGrowthRateYStatic: { default: 0, field: 2, scale: 4, time: 1 },
       unk_ds3_f1_82: { default: 0, field: 2 },
-      particleGrowthRateXMin: { default: 0, field: 2, scale: 5, time: 1 },
-      particleGrowthRateYMin: { default: 0, field: 2, scale: 5, time: 1 },
+      particleGrowthRateXMin: { default: 0, field: 2, scale: 4, time: 1 },
+      particleGrowthRateYMin: { default: 0, field: 2, scale: 4, time: 1 },
       unk_ds3_f1_85: { default: 0, field: 2 },
-      particleGrowthRateXMax: { default: 0, field: 2, scale: 5, time: 1 },
-      particleGrowthRateYMax: { default: 0, field: 2, scale: 5, time: 1 },
+      particleGrowthRateXMax: { default: 0, field: 2, scale: 4, time: 1 },
+      particleGrowthRateYMax: { default: 0, field: 2, scale: 4, time: 1 },
       unk_ds3_f1_88: { default: 0, field: 2 },
-      particleGrowthAccelerationXMin: { default: 0, field: 2, scale: 5, time: 4 },
-      particleGrowthAccelerationYMin: { default: 0, field: 2, scale: 5, time: 4 },
+      particleGrowthAccelerationXMin: { default: 0, field: 2, scale: 4, time: 4 },
+      particleGrowthAccelerationYMin: { default: 0, field: 2, scale: 4, time: 4 },
       unk_ds3_f1_91: { default: 0, field: 2 },
-      particleGrowthAccelerationXMax: { default: 0, field: 2, scale: 5, time: 4 },
-      particleGrowthAccelerationYMax: { default: 0, field: 2, scale: 5, time: 4 },
+      particleGrowthAccelerationXMax: { default: 0, field: 2, scale: 4, time: 4 },
+      particleGrowthAccelerationYMax: { default: 0, field: 2, scale: 4, time: 4 },
       unk_ds3_f1_94: { default: 0, field: 2 },
       rgbMultiplier: { default: 1, field: 2 },
       alphaMultiplier: { default: 1, field: 2 },
@@ -4550,7 +4550,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       particleRandomTurnIntervalMax: { default: 1, field: 1, time: 2 },
       traceParticles: { default: false, field: 0 },
       unk_ds3_f1_149: { default: 1, field: 2 },
-      particleTraceLength: { default: 1, field: 2, scale: 5, time: 2 },
+      particleTraceLength: { default: 1, field: 2, scale: 4, time: 2 },
       traceParticlesThreshold: { default: 0, field: 2 },
       traceParticleHead: { default: false, field: 0 },
       unk_ds3_f1_153: { default: 0, field: 1 },
@@ -4579,19 +4579,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 1 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
       unk_sdt_f2_29: { default: 0, field: 2 },
@@ -4623,8 +4623,8 @@ const ActionData: Record<string, ActionDataEntry> = {
     isParticle: false,
     slotDefault: false,
     props: {
-      width: { default: 1, scale: 5 },
-      height: { default: 1, scale: 5 },
+      width: { default: 1, scale: 4 },
+      height: { default: 1, scale: 4 },
       color1: { default: [1, 1, 1, 1], color: 2 },
       color2: { default: [1, 1, 1, 1], color: 2 },
       color3: { default: [1, 1, 1, 1], color: 1 },
@@ -4741,10 +4741,10 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ac6_f1_60: { default: 0, field: 1 },
       unk_ac6_f1_61: { default: 0, field: 1 },
       unk_ac6_f1_62: { default: 0, field: 1 },
-      particleLengthMin: { default: 1, field: 2, scale: 5 },
-      particleLengthMax: { default: 1, field: 2, scale: 5 },
-      particleWidthMin: { default: 1, field: 2, scale: 5 },
-      particleWidthMax: { default: 1, field: 2, scale: 5 },
+      particleLengthMin: { default: 1, field: 2, scale: 4 },
+      particleLengthMax: { default: 1, field: 2, scale: 4 },
+      particleWidthMin: { default: 1, field: 2, scale: 4 },
+      particleWidthMax: { default: 1, field: 2, scale: 4 },
       unk_ac6_f1_67: { default: 1, field: 2 },
       unk_ac6_f1_68: { default: 1, field: 2 },
       particleDurationMultiplier: { default: 1, field: 2 },
@@ -4793,19 +4793,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ac6_f2_11: { default: 0, field: 1 },
       unk_ac6_f2_12: { default: 0, field: 1 },
       unk_ac6_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ac6_f2_20: { default: 0, field: 1 },
       unk_ac6_f2_21: { default: 0, field: 1 },
       unk_ac6_f2_22: { default: 0, field: 1 },
       unk_ac6_f2_23: { default: 0, field: 1 },
       unk_ac6_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ac6_f2_27: { default: 1, field: 1 },
       unk_ac6_f2_28: { default: 0, field: 1 },
       unk_ac6_f2_29: { default: 0, field: 2 },
@@ -4900,10 +4900,10 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ac6_f1_60: { default: 0, field: 1 },
       unk_ac6_f1_61: { default: 0, field: 1 },
       unk_ac6_f1_62: { default: 0, field: 1 },
-      particleLengthMin: { default: 1, field: 2, scale: 5 },
-      particleLengthMax: { default: 1, field: 2, scale: 5 },
-      particleWidthMin: { default: 1, field: 2, scale: 5 },
-      particleWidthMax: { default: 1, field: 2, scale: 5 },
+      particleLengthMin: { default: 1, field: 2, scale: 4 },
+      particleLengthMax: { default: 1, field: 2, scale: 4 },
+      particleWidthMin: { default: 1, field: 2, scale: 4 },
+      particleWidthMax: { default: 1, field: 2, scale: 4 },
       unk_ac6_f1_67: { default: 1, field: 2 },
       unk_ac6_f1_68: { default: 1, field: 2 },
       particleDurationMultiplier: { default: 1, field: 2 },
@@ -4952,19 +4952,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ac6_f2_11: { default: 0, field: 1 },
       unk_ac6_f2_12: { default: 0, field: 1 },
       unk_ac6_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ac6_f2_20: { default: 0, field: 1 },
       unk_ac6_f2_21: { default: 0, field: 1 },
       unk_ac6_f2_22: { default: 0, field: 1 },
       unk_ac6_f2_23: { default: 0, field: 1 },
       unk_ac6_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ac6_f2_27: { default: 1, field: 1 },
       unk_ac6_f2_28: { default: 0, field: 1 },
       unk_ac6_f2_29: { default: 0, field: 2 },
@@ -4996,7 +4996,7 @@ const ActionData: Record<string, ActionDataEntry> = {
     props: {
       texture: { default: 1, field: 1, resource: 0, textureType: 'a' },
       blendMode: { default: BlendMode.Normal, field: 1 },
-      width: { default: 1, scale: 5 },
+      width: { default: 1, scale: 4 },
       widthMultiplier: { default: 1 },
       startFadeEndpoint: { default: 0 },
       endFadeEndpoint: { default: 0 },
@@ -5020,7 +5020,7 @@ const ActionData: Record<string, ActionDataEntry> = {
       orientation: { default: TracerOrientationMode.LocalZ, field: 1 },
       normalMap: { default: 0, field: 1, resource: 0, textureType: 'n' },
       segmentInterval: { default: 0, field: 2, time: 2 },
-      segmentDuration: { default: 1, field: 2, scale: 5, time: 5 },
+      segmentDuration: { default: 1, field: 2, scale: 4, time: 5 },
       concurrentSegments: { default: 100, field: 1 },
       segmentSubdivision: { default: 0, field: 1 },
       unk_ds3_f1_8: { default: 0, field: 2 },
@@ -5050,22 +5050,22 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_ds3_f2_11: { default: 0, field: 1 },
       unk_ds3_f2_12: { default: 0, field: 1 },
       unk_ds3_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_ds3_f2_20: { default: 0, field: 2 },
       unk_ds3_f2_21: { default: 0, field: 1 },
       unk_ds3_f2_22: { default: 0, field: 1 },
       unk_ds3_f2_23: { default: 0, field: 1 },
       unk_ds3_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_ds3_f2_27: { default: 1, field: 1 },
       unk_ds3_f2_28: { default: 0, field: 1 },
-      unk_ds3_f2_29: { default: 0, field: 2, scale: 3 },
+      unk_ds3_f2_29: { default: 0, field: 2, scale: 2 },
       shadowDarkness: { default: 0, field: 2 },
       unk_sdt_f2_31: { default: 0, field: 1 },
       unk_sdt_f2_32: { default: false, field: 0 },
@@ -5112,8 +5112,8 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: false,
     props: {
       texture: { default: 50004, field: 1, resource: 0, textureType: 'd' },
-      depth: { default: 1, field: 2, scale: 5 },
-      size: { default: 1, field: 2, scale: 5 },
+      depth: { default: 1, field: 2, scale: 4 },
+      size: { default: 1, field: 2, scale: 4 },
       descent: { default: 0.15, field: 2, time: 2 },
       duration: { default: 0.15, field: 2, time: 2 },
     },
@@ -5130,17 +5130,17 @@ const ActionData: Record<string, ActionDataEntry> = {
     isParticle: false,
     slotDefault: false,
     props: {
-      layer1Width: { default: 1, scale: 5 },
-      layer1Height: { default: 1, scale: 5 },
+      layer1Width: { default: 1, scale: 4 },
+      layer1Height: { default: 1, scale: 4 },
       layer1Color: { default: [1, 1, 1, 1], color: 1 },
-      layer2Width: { default: 1, scale: 5 },
-      layer2Height: { default: 1, scale: 5 },
+      layer2Width: { default: 1, scale: 4 },
+      layer2Height: { default: 1, scale: 4 },
       layer2Color: { default: [1, 1, 1, 1], color: 1 },
-      layer3Width: { default: 1, scale: 5 },
-      layer3Height: { default: 1, scale: 5 },
+      layer3Width: { default: 1, scale: 4 },
+      layer3Height: { default: 1, scale: 4 },
       layer3Color: { default: [1, 1, 1, 1], color: 1 },
-      layer4Width: { default: 1, scale: 5 },
-      layer4Height: { default: 1, scale: 5 },
+      layer4Width: { default: 1, scale: 4 },
+      layer4Height: { default: 1, scale: 4 },
       layer4Color: { default: [1, 1, 1, 1], color: 1 },
       layer1: { default: 1, field: 1, resource: 0, textureType: 'a' },
       layer2: { default: 0, field: 1, resource: 0, textureType: 'a' },
@@ -5257,9 +5257,9 @@ const ActionData: Record<string, ActionDataEntry> = {
     slotDefault: false,
     props: {
       model: { default: 80201, resource: 1 },
-      sizeX: { default: 1, scale: 5 },
-      sizeY: { default: 1, scale: 5 },
-      sizeZ: { default: 1, scale: 5 },
+      sizeX: { default: 1, scale: 4 },
+      sizeY: { default: 1, scale: 4 },
+      sizeZ: { default: 1, scale: 4 },
       rotationX: { default: 0 },
       rotationY: { default: 0 },
       rotationZ: { default: 0 },
@@ -5341,19 +5341,19 @@ const ActionData: Record<string, ActionDataEntry> = {
       unk_er_f2_11: { default: 0, field: 1 },
       unk_er_f2_12: { default: 0, field: 1 },
       unk_er_f2_13: { default: 0, field: 1 },
-      minFadeDistance: { default: -1, field: 2, scale: 4 },
-      minDistance: { default: -1, field: 2, scale: 4 },
-      maxFadeDistance: { default: -1, field: 2, scale: 4 },
-      maxDistance: { default: -1, field: 2, scale: 4 },
-      minDistanceThreshold: { default: -1, field: 2, scale: 4 },
-      maxDistanceThreshold: { default: -1, field: 2, scale: 4 },
+      minFadeDistance: { default: -1, field: 2, scale: 3 },
+      minDistance: { default: -1, field: 2, scale: 3 },
+      maxFadeDistance: { default: -1, field: 2, scale: 3 },
+      maxDistance: { default: -1, field: 2, scale: 3 },
+      minDistanceThreshold: { default: -1, field: 2, scale: 3 },
+      maxDistanceThreshold: { default: -1, field: 2, scale: 3 },
       unk_er_f2_20: { default: 0, field: 1 },
       unk_er_f2_21: { default: 0, field: 1 },
       unk_er_f2_22: { default: 0, field: 1 },
       unk_er_f2_23: { default: 0, field: 1 },
       unk_er_f2_24: { default: 0, field: 1 },
       unkDepthBlend1: { default: 1, field: 2 },
-      unkDepthBlend2: { default: 0, field: 2, scale: 3 },
+      unkDepthBlend2: { default: 0, field: 2, scale: 2 },
       unk_er_f2_27: { default: 0, field: 1 },
       unk_er_f2_28: { default: 1, field: 1 },
       unk_er_f2_29: { default: 0, field: 1 },
@@ -5784,7 +5784,7 @@ const ActionData: Record<string, ActionDataEntry> = {
     props: {
       rateOfTime: { default: 1, field: 2 },
       limitViewDistance: { default: false, field: 0 },
-      maxViewDistance: { default: 0, field: 2, scale: 3 },
+      maxViewDistance: { default: 0, field: 2, scale: 2 },
       unk_ds3_f1_2: { default: 0, field: 2 },
       unk_ds3_f1_3: { default: 0, field: 2 },
       unk_ds3_f1_4: { default: 0, field: 2 },
@@ -5820,10 +5820,10 @@ const ActionData: Record<string, ActionDataEntry> = {
       specularColor: { default: [1, 1, 1, 1], color: 1 },
       diffuseMultiplier: { default: 1 },
       specularMultiplier: { default: 1 },
-      near: { default: 0.01, scale: 5 },
-      far: { default: 50, scale: 5 },
-      radiusX: { default: 50, scale: 5 },
-      radiusY: { default: 50, scale: 5 },
+      near: { default: 0.01, scale: 4 },
+      far: { default: 50, scale: 4 },
+      radiusX: { default: 50, scale: 4 },
+      radiusY: { default: 50, scale: 4 },
       unk_ds3_p1_6: { default: 1 },
       unk_ds3_p1_7: { default: 1 },
       unk_sdt_p1_10: { default: 1 },
@@ -10894,7 +10894,7 @@ abstract class Node {
    * @param options.recurse Controls whether or not the scaling should be applied to
    * all descendant nodes. Defaults to true.
    */
-  scale(factor: number, options: { recurse?: boolean, mode?: ScalingMode } = {}) {
+  scale(factor: number, options: ScaleOptions & { recurse?: boolean } = defaultScaleOptions) {
     if (this instanceof NodeWithConfigs || this instanceof GenericNode) {
       for (const config of this.configs) {
         config.scale(factor, options)
@@ -11970,7 +11970,7 @@ class NodeConfig implements IConfig {
     )
   }
 
-  scale(factor: number, options?: { mode?: ScalingMode }) {
+  scale(factor: number, options: ScaleOptions = defaultScaleOptions) {
     for (const action of this.walkActions()) if (action instanceof DataAction) {
       action.scale(factor, options)
     }
@@ -12040,8 +12040,8 @@ class LevelsOfDetailConfig implements IConfig {
     )
   }
 
-  scale(factor: number, options: { mode?: ScalingMode } = {}) {
-    if (options.mode === ScalingMode.All) {
+  scale(factor: number, options: ScaleOptions = defaultScaleOptions) {
+    if (options.scaleStructural && options.scaleViewDistance) {
       this.thresholds = this.thresholds.slice(0, 5).map(t => t * factor)
     }
     return this
@@ -12242,7 +12242,7 @@ class BasicConfig implements IConfig {
     })
   }
 
-  scale(factor: number, options: { mode?: ScalingMode } = {}) {
+  scale(factor: number, options: ScaleOptions = defaultScaleOptions) {
     for (const action of this.walkActions()) if (action instanceof DataAction) {
       action.scale(factor, options)
     }
@@ -12545,7 +12545,7 @@ class NodeEmitterConfig implements IConfig {
     })
   }
 
-  scale(factor: number, options: { mode?: ScalingMode } = {}) {
+  scale(factor: number, options: ScaleOptions = defaultScaleOptions) {
     for (const action of this.walkActions()) if (action instanceof DataAction) {
       action.scale(factor, options)
     }
@@ -12867,23 +12867,23 @@ class DataAction implements IAction {
    * used with actions that have scaling properties.
    * @param factor The factor to scale by.
    */
-  scale(factor: number, options: { mode?: ScalingMode } = {}) {
+  scale(factor: number, options: ScaleOptions = defaultScaleOptions) {
     if ('props' in this.$data) {
       for (const [k, v] of Object.entries(this.$data.props)) {
         if ('scale' in v) {
           switch (true) {
             case v.scale === ScaleCondition.True
-              && options.mode !== ScalingMode.InstancesOnly
-              && options.mode !== ScalingMode.ParticleModifierOnly:
+              && options.scaleStructural:
 
             case v.scale === ScaleCondition.InstanceSize
-              && options.mode !== ScalingMode.ParticleModifierOnly:
+              && !options.scaleParticleModifier:
 
             case v.scale === ScaleCondition.Distance
-              && options.mode === ScalingMode.All:
+              && options.scaleStructural
+              && options.scaleViewDistance:
 
             case v.scale === ScaleCondition.ParticleModifier
-              && options.mode === ScalingMode.ParticleModifierOnly:
+              && options.scaleParticleModifier:
 
               {
                 this[k] = anyValueMult(factor, this[k])
@@ -12891,12 +12891,9 @@ class DataAction implements IAction {
               }
 
 
-            case v.scale === ScaleCondition.IfNotMinusOne
-              && options.mode !== ScalingMode.ParticleModifierOnly
-              && this[k] !== -1:
-
             case v.scale === ScaleCondition.DistanceIfNotMinusOne
-              && options.mode === ScalingMode.All
+              && options.scaleStructural
+              && options.scaleViewDistance
               && this[k] !== -1:
 
               {